+++ crates/voicevox_core/src/engine/talk.rs
@@ -8,6 +8,9 @@ pub(crate) mod text_analyzer;
 pub(crate) mod user_dict;
 
 pub use self::audio_query::{AccentPhrase, AudioQuery, Mora};
+pub(crate) use self::audio_query::{
+    LengthedPhoneme, ValidatedAccentPhrase, ValidatedAudioQuery, ValidatedMora,
+};
 pub(crate) use self::full_context_label::extract_full_context_label;
 pub(crate) use self::interpret_query::{initial_process, split_mora, DecoderFeature};
 pub(crate) use self::kana_parser::{create_kana, parse_kana, KanaParseError};