+++ crates/voicevox_core_java_api/lib/src/main/java/jp/hiroshiba/voicevoxcore/exceptions/InvalidQueryException.java
@@ -1,0 +1,12 @@
+package jp.hiroshiba.voicevoxcore.exceptions;
+
+/** AudioQuery、もしくはその一部が不正。 */
+public class InvalidQueryException extends IllegalArgumentException {
+  public InvalidQueryException(String message) {
+    super(message);
+  }
+
+  public InvalidQueryException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}