+++ crates/voicevox_core_java_api/src/common.rs
@@ -93,6 +93,7 @@ where
                             WordNotFound,
                             UseUserDict,
                             InvalidWord,
+                            InvalidQuery,
                         );
 
                         let mut sources =